# src/swagger.json
{
	"components": {
		"examples": {},
		"headers": {},
		"parameters": {},
		"requestBodies": {},
		"responses": {},
		"schemas": {
			"SignInDto": {
				"properties": {
					"email": {
						"type": "string"
					},
					"password": {
						"type": "string"
					}
				},
				"required": [
					"email",
					"password"
				],
				"type": "object",
				"additionalProperties": false
			},
			"BookDetailDto": {
				"properties": {
					"room_id": {
						"type": "number",
						"format": "double"
					},
					"from_date": {
						"type": "string",
						"format": "date-time"
					},
					"to_date": {
						"type": "string",
						"format": "date-time"
					}
				},
				"required": [
					"room_id",
					"from_date",
					"to_date"
				],
				"type": "object",
				"additionalProperties": false
			},
			"CreateBookingDto": {
				"properties": {
					"data": {
						"items": {
							"$ref": "#/components/schemas/BookDetailDto"
						},
						"type": "array"
					}
				},
				"required": [
					"data"
				],
				"type": "object",
				"additionalProperties": false
			},
			"UpdateBookingDto": {
				"properties": {
					"from_date": {
						"type": "string",
						"format": "date-time"
					},
					"to_date": {
						"type": "string",
						"format": "date-time"
					}
				},
				"required": [
					"from_date",
					"to_date"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ImagesDto": {
				"properties": {
					"path": {
						"type": "string"
					}
				},
				"required": [
					"path"
				],
				"type": "object",
				"additionalProperties": false
			},
			"CreateRoomDto": {
				"properties": {
					"name": {
						"type": "string"
					},
					"type": {
						"type": "string"
					},
					"description": {
						"type": "string"
					},
					"images": {
						"items": {
							"$ref": "#/components/schemas/ImagesDto"
						},
						"type": "array"
					},
					"price": {
						"type": "number",
						"format": "double"
					},
					"quantity": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"name",
					"type",
					"description",
					"images",
					"price",
					"quantity"
				],
				"type": "object",
				"additionalProperties": false
			},
			"UpdateRoomDto": {
				"properties": {
					"name": {
						"type": "string"
					},
					"type": {
						"type": "string"
					},
					"description": {
						"type": "string"
					},
					"images": {
						"items": {
							"$ref": "#/components/schemas/ImagesDto"
						},
						"type": "array"
					},
					"price": {
						"type": "number",
						"format": "double"
					},
					"quantity": {
						"type": "number",
						"format": "double"
					}
				},
				"type": "object",
				"additionalProperties": false
			}
		},
		"securitySchemes": {
			"bearer": {
				"type": "apiKey",
				"in": "header",
				"name": "Authorization"
			}
		}
	},
	"info": {
		"title": "care_be",
		"version": "1.0.0",
		"license": {
			"name": "ISC"
		},
		"contact": {}
	},
	"openapi": "3.0.0",
	"paths": {
		"/auth/sign-in": {
			"post": {
				"operationId": "SignIn",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {}
							}
						}
					}
				},
				"tags": [
					"Auth"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/SignInDto"
							}
						}
					}
				}
			}
		},
		"/auth/sign-up": {
			"post": {
				"operationId": "SignUp",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"success": {
											"type": "boolean"
										},
										"data": {}
									},
									"required": [
										"success",
										"data"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"tags": [
					"Auth"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/SignInDto"
							}
						}
					}
				}
			}
		},
		"/bookings": {
			"post": {
				"operationId": "Create",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {}
							}
						}
					}
				},
				"tags": [
					"Bookings"
				],
				"security": [
					{
						"bearer": [
							"CUSTOMER"
						]
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/CreateBookingDto"
							}
						}
					}
				}
			},
			"get": {
				"operationId": "GetAll",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"success": {
											"type": "boolean"
										},
										"data": {}
									},
									"required": [
										"success",
										"data"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"tags": [
					"Bookings"
				],
				"security": [
					{
						"bearer": [
							"CUSTOMER"
						]
					}
				],
				"parameters": [
					{
						"in": "query",
						"name": "page",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "limit",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "status",
						"required": false,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/bookings/{id}": {
			"get": {
				"operationId": "GetOne",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"success": {
											"type": "boolean"
										},
										"data": {}
									},
									"required": [
										"success",
										"data"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"tags": [
					"Bookings"
				],
				"security": [
					{
						"bearer": [
							"CUSTOMER"
						]
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			},
			"put": {
				"operationId": "Update",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"success": {
											"type": "boolean"
										},
										"data": {}
									},
									"required": [
										"success",
										"data"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"tags": [
					"Bookings"
				],
				"security": [
					{
						"bearer": [
							"CUSTOMER"
						]
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/UpdateBookingDto"
							}
						}
					}
				}
			},
			"delete": {
				"operationId": "CancelBooking",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"success": {
											"type": "boolean"
										},
										"data": {}
									},
									"required": [
										"success",
										"data"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"tags": [
					"Bookings"
				],
				"security": [
					{
						"bearer": [
							"CUSTOMER"
						]
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
		"/rooms": {
			"post": {
				"operationId": "Create",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {}
							}
						}
					}
				},
				"tags": [
					"Rooms"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/CreateRoomDto"
							}
						}
					}
				}
			},
			"get": {
				"operationId": "GetAll",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"success": {
											"type": "boolean"
										},
										"data": {}
									},
									"required": [
										"success",
										"data"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"tags": [
					"Rooms"
				],
				"security": [],
				"parameters": [
					{
						"in": "query",
						"name": "page",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "limit",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
		"/rooms/upload-images": {
			"post": {
				"operationId": "UploadImages",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"success": {
											"type": "boolean"
										},
										"data": {}
									},
									"required": [
										"success",
										"data"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"tags": [
					"Rooms"
				],
				"security": [],
				"parameters": [],
				"consumes": [
					"multipart/form-data"
				],
				"requestBody": {
					"required": true,
					"content": {
						"multipart/form-data": {
							"schema": {
								"type": "object",
								"properties": {
									"images": {
										"items": {
											"type": "string",
											"format": "binary"
										},
										"type": "array"
									}
								},
								"required": [
									"images"
								]
							}
						}
					}
				}
			}
		},
		"/rooms/available": {
			"get": {
				"operationId": "GetAllRoomsAvailable",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"success": {
											"type": "boolean"
										},
										"data": {}
									},
									"required": [
										"success",
										"data"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"tags": [
					"Rooms"
				],
				"security": [],
				"parameters": [
					{
						"in": "query",
						"name": "date",
						"required": false,
						"schema": {
							"format": "date-time",
							"type": "string"
						}
					}
				]
			}
		},
		"/rooms/{id}": {
			"get": {
				"operationId": "GetOne",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"success": {
											"type": "boolean"
										},
										"data": {}
									},
									"required": [
										"success",
										"data"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"tags": [
					"Rooms"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			},
			"put": {
				"operationId": "Update",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"success": {
											"type": "boolean"
										},
										"data": {}
									},
									"required": [
										"success",
										"data"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"tags": [
					"Rooms"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/UpdateRoomDto"
							}
						}
					}
				}
			},
			"delete": {
				"operationId": "Delete",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"success": {
											"type": "boolean"
										},
										"data": {}
									},
									"required": [
										"success",
										"data"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"tags": [
					"Rooms"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		}
	},
	"servers": [
		{
			"url": "/api/v1"
		}
	]
}